+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @grid-table/basic
 
+## 0.6.44
+
+### Patch Changes
+
+- ColumnResizeCallBack
+- Updated dependencies
+  - @grid-table/core@0.6.44
+
 ## 0.6.43
 
 ### Patch Changes